+++ PiwikTracker/URLSessionDispatcher.swift
@@ -105,6 +105,10 @@ fileprivate extension Event {
                 URLQueryItem(name: "m", value: DateFormatter.minuteDateFormatter.string(from: date)),
                 URLQueryItem(name: "s", value: DateFormatter.secondsDateFormatter.string(from: date)),
                 
+                
+                //screen resolution
+                URLQueryItem(name: "res", value:String(format: "%1.0fx%1.0f", screenResolution.width, screenResolution.height)),
+
                 URLQueryItem(name: "e_c", value: eventCategory),
                 URLQueryItem(name: "e_a", value: eventAction),
                 URLQueryItem(name: "e_n", value: eventName),