+++ Example/iOS Example/MenuViewController.swift
@@ -1,9 +1,0 @@
-import UIKit
-import PiwikTracker
-
-class MenuViewController: UITableViewController {
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        Tracker.shared?.track(view: ["menu"])
-    }
-}