+++ Example/iOS Example/ScreenViewController.swift
@@ -1,9 +1,0 @@
-import UIKit
-import PiwikTracker
-
-class ScreenViewController: UIViewController {
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        Tracker.shared?.track(view: ["menu","screen view"])
-    }
-}