+++ Example/iOS Example/SecondScreenViewController.swift
@@ -1,9 +1,0 @@
-import UIKit
-import PiwikTracker
-
-class SecondScreenViewController: UIViewController {
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        Tracker.shared?.track(view: ["menu","screen view", "screen view 2"])
-    }
-}