+++ Example/ios/iOS Example/MenuViewController.swift
@@ -1,0 +1,9 @@
+import UIKit
+import PiwikTracker
+
+class MenuViewController: UITableViewController {
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        PiwikTracker.shared?.track(view: ["menu"])
+    }
+}