+++ Example/ios/iOS Example/ScreenViewController.swift
@@ -1,0 +1,9 @@
+import UIKit
+import PiwikTracker
+
+class ScreenViewController: UIViewController {
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        PiwikTracker.shared?.track(view: ["menu","screen view"])
+    }
+}