+++ Example/macos/macos/AppDelegate.swift
@@ -1,0 +1,26 @@
+//
+//  AppDelegate.swift
+//  macos
+//
+//  Created by Cornelius Horstmann on 07.05.17.
+//  Copyright © 2017 Piwik. All rights reserved.
+//
+
+import Cocoa
+
+@NSApplicationMain
+class AppDelegate: NSObject, NSApplicationDelegate {
+
+
+
+    func applicationDidFinishLaunching(_ aNotification: Notification) {
+        // Insert code here to initialize your application
+    }
+
+    func applicationWillTerminate(_ aNotification: Notification) {
+        // Insert code here to tear down your application
+    }
+
+
+}
+