+++ PiwikTracker/Dispatcher.swift
@@ -6,7 +6,7 @@ public protocol Dispatcher {
     
     var userAgent: String? { get }
     
-    func send(event: Event, success: @escaping ()->(), failure: @escaping (_ shouldContinue: Bool)->())
+    func send(event: Event, success: @escaping ()->(), failure: @escaping (_ error: Error)->())
     
-    func send(events: [Event], success: @escaping ()->(), failure: @escaping (_ shouldContinue: Bool)->())
+    func send(events: [Event], success: @escaping ()->(), failure: @escaping (_ error: Error)->())
 }