+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## Unreleased
+* **feature** Transmitting the Screen resolution to the Piwik Backend. [#149](https://github.com/piwik/piwik-sdk-ios/issues/149) (by @akshaykolte)
+* **feature** Added a Logger that can log messages in different levels. [#147](https://github.com/piwik/piwik-sdk-ios/issues/147)
+* **feature** Added macOS and tvOS compatibility. [#134](https://github.com/piwik/piwik-sdk-ios/issues/134)
+
+## 4.0.0
+* **feature** Renamed the Tracker to PiwikTracker. [#146](https://github.com/piwik/piwik-sdk-ios/issues/146)
+* **feature** Added isOptedOut parameter to the PiwikTracker. [#124](https://github.com/piwik/piwik-sdk-ios/issues/124)
+
+## 4.0.0-beta2
+* **feature** Added the possibility to set the url for a sreen view event. [#92](https://github.com/piwik/piwik-sdk-ios/issues/92)
+* **feature** Added the functionality to start new sessions. [#136](https://github.com/piwik/piwik-sdk-ios/issues/136)
+* **fixed** The value of an event got wronly encoded when dispatching. [#140](https://github.com/piwik/piwik-sdk-ios/pull/140)
+* **fixed** Fixed an issue where tracking an event wasn’t possible from Objective-C code. [#142](https://github.com/piwik/piwik-sdk-ios/issues/142)
+
+## 4.0.0-beta1
+* no changes
+
 ## 4.0.0-alpha1
 * **feature** Started rewrite in Swift
 * **feature** Tracking Screen Views